requirements.txt
streamlit==1.32.0
openai==1.12.0
python-dotenv==1.0.1
requests==2.31.0
pandas==2.2.1
asyncio==3.4.3
streamlit-folium==0.18.0
streamlit-drawable-canvas==0.9.3
geojson==3.1.0
shapely==2.0.3
folium==0.15.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
werkzeug==3.0.1